[IMP] mail,im_livechat: no canned reply in mail
Canned response should only enabled in livechat thread.
task-3208438
